Rename error fallback file
