+++ apps/client-react/src/App.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Navigate, RouterProvider, createBrowserRouter } from "react-router-dom";
 
-import GeneralErrorFallback from "~/components/general-error-fallback";
+import GeneralErrorFallback from "~/components/error-fallback";
 import ErrorBoundary from "~/components/ui/error-boundary";
 import ProjectCreateRoute from "~/views/project-create-route";
 import ProjectEditRoute from "~/views/project-edit-route";
